Make subtitle of readme look better
Instead of making the subtitle be a header like the title,
just make it normal text.
This is done because when the subtitle is a header, it is not attached
to the main title. A dividing line appears between the subtitle and
title. And so there is little point of trying to make the subtitle be
part of the title.
Instead, make the subtitle be it's own thing/paragraph,
separate from the title.